requirements.txt
Flask==2.0.2
Flask-MySQLdb==0.2.0
python-dotenv==0.19.0
Flask-Cors==3.0.10
mysql-connector-python==8.0.27